Add gcdoe certification (#163)
* Add Google Cloud DevOps Engineer to certificate list

A new entry for the "Google Cloud Certification - Professional DevOps Engineer" is added to the list in the Certificate component. This entry will help to display this specific certification correctly in the UI.

* merge main

* regenerate yarn.lock

* fix: eslint

* Rename ESLint configuration file and update its structure

The eslintrc.js file has been renamed to eslint.config.js. Furthermore, the structure and settings within the ESLint configuration have been updated to use ES modules, leading to better compatibility and easier configuration. The package.json has also been modified to denote the project type as "module".

* make config files module